+++ src/test/ExSln2/TranspileRunner/Program.cs
@@ -11,7 +11,7 @@ string slnName = "ExSln2.sln";
 string outDir = slnDir + "/c99/gen";
 string projectName = "LedBlinker";
 
-Environment.SetEnvironmentVariable(CTranspiler.ENV_VAR_TRANSPILER_DEBUG_TYPE, nameof(VtableReturnsObjPointerEx));
+Environment.SetEnvironmentVariable(CTranspiler.ENV_VAR_TRANSPILER_DEBUG_TYPE, nameof(FuncPtrEx2));
 
 Console.WriteLine("Transpiling " + projectName + " fin/C# project...");