+++ src/test/ExSln2/c99/gen/hal_ConstantsClassEx.c
@@ -4,8 +4,14 @@
 
 
 #include "hal_ConstantsClassEx.h"
+#include <string.h>
 
 
+void hal_ConstantsClassEx_ctor(hal_ConstantsClassEx * self)
+{
+    memset(self, 0, sizeof(*self));
+    self->my_div = 2;
+}
 
 uint32_t hal_ConstantsClassEx_calc_as_instance(hal_ConstantsClassEx * self, uint32_t cycles)
 {