+++ src/test/ExSln2/c99/gen/issue58_Bike.c
@@ -7,10 +7,8 @@
 #include <string.h>
 
 
-
 void issue58_Bike_ctor(issue58_Bike * self)
 {
     memset(self, 0, sizeof(*self));
     self->_speed = 0;
-    // needed for now
 }